backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.26.0
asyncpg==0.29.0
python-telegram-bot==20.8
requests==2.31.0
